feat(watch) #379: prev_kv + progress heartbeat
## Production code

- WatchEvent: add `prev_value: Bytes` field; add `WatchEventType::Progress` variant
- WatchRegistry: `register()` / `register_prefix()` take `prev_kv: bool`; expose
  `prev_kv_watcher_count_arc()` for shared atomic counter
- DefaultStateMachineHandler: when `prev_kv_watcher_count > 0`, read old values from
  state machine before `apply_chunk` and populate `prev_value` in broadcast events
- WatchDispatcher: `heartbeat_interval_ms` param; tokio interval with ±10% jitter
  broadcasts `Progress` events to all watchers; `broadcast_progress()` fan-out
  iterates registry maps directly (key-routing would miss watchers on empty key)
- EmbeddedClient: add `watch_with_options(key, prev_kv)` and
  `watch_prefix_with_options(prefix, prev_kv)` as explicit prev_kv entry points
- Proto: add `prev_value` field to `WatchResponse`; add `prev_kv` field to `WatchRequest`
- Config: add `heartbeat_interval_ms: u64` to `WatchConfig`
- NodeBuilder: wire `prev_kv_watcher_count` Arc into `DefaultStateMachineHandler::new()`

## Tests (RED → GREEN)

- manager_test: 6 new tests — prev_kv isolation per watcher, watcher count tracking,
  progress heartbeat delivery, heartbeat disabled produces no Progress events
- default_state_machine_handler_test: 3 new tests — skip get when count=0, read and
  broadcast prev_value when count=1, stop reading after count drops to zero
- grpc_client_test: add `prev_value` field to WatchResponse fixture
- embedded_client_test: guard both test mods with `#[cfg(all(test, feature="rocksdb"))]`

## API compatibility fixes (callers broken by new fields / signatures)

- benches/state_machine.rs: `prev_value`, `register(_, false)`, WatchDispatcher 5-arg
- tests/embedded_client: `WatchEventType::X as i32` → `WatchEventType::X`
- tests/watch_events_embedded: same
- examples/service-discovery-embedded: same + explicit Canceled arm
- examples/service-discovery-standalone: same + Progress arm for exhaustive match
- integration/mod.rs: DefaultStateMachineHandler::new() 7th arg

## Docs

- watch-feature.md: replace "No prev_kv / No heartbeat (planned)" with feature docs;
  add `heartbeat_interval_ms` config entry; fix `WatchEventType::CANCELED` casing (×3);
  fix `Canceled as i32` in reconnection pattern; add Progress to Best Practices
- embedded_client.rs: full rustdoc for `watch_with_options` and
  `watch_prefix_with_options` — prev_kv semantics, performance note, empty-prev_value
  cases, code examples
- service-discovery-embedded: Demo 3 — prev_kv audit log pattern (put/delete/canceled/progress)

+++ d-engine-server/tests/embedded_client/embedded_client_operations.rs
@@ -724,11 +724,7 @@ async fn test_watch_receives_put_event() {
         .expect("timed out waiting for Put watch event")
         .expect("watch channel should not be closed");
 
-    assert_eq!(
-        event.event_type,
-        WatchEventType::Put as i32,
-        "expected Put event"
-    );
+    assert_eq!(event.event_type, WatchEventType::Put, "expected Put event");
     assert_eq!(event.key, Bytes::from("w-key"));
     assert_eq!(event.value, Bytes::from("w-val"));
 }
@@ -755,7 +751,7 @@ async fn test_watch_receives_delete_event() {
 
     assert_eq!(
         event.event_type,
-        WatchEventType::Delete as i32,
+        WatchEventType::Delete,
         "expected Delete event"
     );
     assert_eq!(event.key, Bytes::from("d-key"));
